not refering to window for server-side
