+++ viewModel.js
@@ -1,34 +1,7 @@
 var domComponent = require('./domComponent');
 var rendering = require('./rendering');
 var plastiqMeta = require('./meta');
-
-var Widgets;
-if (typeof window.Set === 'function') {
-  Widgets = window.Set;
-} else {
-  Widgets = function() {
-    this.widgets = [];
-  };
-
-  Widgets.prototype.add = function(widget) {
-    if (this.widgets.indexOf(widget) == -1) {
-      this.widgets.push(widget);
-    }
-  };
-
-  Widgets.prototype.delete = function(widget) {
-    var i = this.widgets.indexOf(widget);
-    if (i !== -1) {
-      this.widgets.splice(i, 1);
-    }
-  };
-
-  Widgets.prototype.forEach = function(fn) {
-    for(var n = 0; n < this.widgets.length; n++) {
-      fn(this.widgets[n]);
-    }
-  };
-}
+var Widgets = require('./widgets');
 
 function rerenderViewModel() {
   var meta = plastiqMeta(this);
